+++ src/components/HeaderHalva/index.ts
@@ -1,1 +1,0 @@
-export { default as HeaderHalva } from "./HeaderHalva";