+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./base";
 export * from "./Icons";
 export * from "./Notification";
-export * from "./HeaderHalva";
+export * from "./halva";